+++ src/common/types.ts
@@ -74,6 +74,8 @@ export type GetInputFunction = (
   options?: RequestOptions,
 ) => Promise<ElicitResult>;
 
+export type GetEnvFn = (key: string, client?: Client) => string | undefined;
+
 export interface Client {
   /** Human-readable name for the client - usually the product name */
   name: string;
@@ -92,6 +94,14 @@ export interface Client {
     [key: string]: ZodType;
   }>;
   /**
+   * Zod schema defining authentication fields for this client
+   * Field names must use snake case to ensure they are mapped to environment variables and HTTP headers correctly.
+   * e.g., `authenticationFields.my_property` would refer to the http header `Tool-My-Property`
+   */
+  authenticationFields: ZodObject<{
+    [key: string]: ZodType;
+  }>;
+  /**
    * Configure the client with the given server and configuration
    */
   configure: (server: SmartBearMcpServer, config: any) => Promise<void>;
@@ -103,9 +113,8 @@ export interface Client {
   registerResources?(register: RegisterResourceFunction): Promise<void>;
   registerPrompts?(register: RegisterPromptFunction): Promise<void>;
   /**
-   * Optional method to retrieve the authentication token for the current request context.
-   * This is used for request-level authentication where the token might change per request.
+   * Whether the client is currently authorized to make API requests.
    */
-  getAuthToken?(): string | null;
+  hasAuth(): boolean;
   cleanupSession?(mcpSessionId: string): Promise<void>;
 }